changes requested in the latest PR addressed and the TPA and KPA have been seperated into their own modules
