+++ src/thorlabs_cube/driver/message.py
@@ -115,18 +115,26 @@ class MGMSG(Enum):
     PZ_SET_TPZ_IOSETTINGS = 0x07D4
     PZ_REQ_TPZ_IOSETTINGS = 0x07D5
     PZ_GET_TPZ_IOSETTINGS = 0x07D6
-
     QUAD_SET_PARAMS = 0x0870
     QUAD_REQ_PARAMS = 0x0871
     QUAD_GET_PARAMS = 0x0872
-
-    QUAD_OPER_MODE_SUB_ID = 0x07
-    QUAD_DISP_SETTINGS_SUB_ID = 0x0B
-
     QUAD_SET_EEPROM_PARAMS = 0x0875
     QUAD_REQ_STATUSUPDATE = 0x0880
     QUAD_GET_STATUSUPDATE = 0x0881
     QUAD_ACK_STATUSUPDATE = 0x0882
+
+
+class SUBMSG(Enum):
+    QUAD_LOOP_PARAMS_SUB_ID: int = 0x01
+    QUAD_READINGS_SUB_ID: int = 0x03
+    QUAD_POSITION_DEMAND_PARAMS_SUB_ID: int = 0x05
+    QUAD_OPER_MODE_SUB_ID: int = 0x07
+    QUAD_STATUS_BITS_SUB_ID: int = 0x09
+    QUAD_DISP_SETTINGS_SUB_ID: int = 0x0B
+    QUAD_POSITION_DEMAND_OUTPUTS_SUB_ID: int = 0x0D
+    QUAD_LOOP_PARAMS2_SUB_ID: int = 0x0E
+    QUAD_KPA_TRIGIO_SUB_ID: int = 0x0F
+    QUAD_KPA_DIGOPS_SUB_ID: int = 0x10
 
 
 class Direction:
